chore(ee): move libs/analytics-warehouse under libs/ee/
The analytics warehouse is the data layer behind the EE-gated cockpit
(tier-policy: Teams cloud / Enterprise / trial). Moving it under
`libs/ee/` aligns its filesystem location with its license coverage,
per the EE audit (item #2).

- git mv libs/analytics-warehouse libs/ee/analytics-warehouse (20 files,
  detected as renames)
- update @libs/analytics-warehouse → @libs/ee/analytics-warehouse in 11
  consumers (cockpit services + module, api/worker/analytics-cli entry
  points, ingestion + classifier crons, backfill script, 2 specs)
- update package.json analytics:migration:* script paths (source + dist)
- update docker/prod-entrypoint.sh dist ormconfig path
- update path comment in scripts/analytics/ensure-schema.cli.ts

Yarn script names are unchanged, so infra (ECS task defs invoking
`yarn analytics:*:prod`) is unaffected — the dist path change is
internal to the entrypoint shell script.
