+++ libs/ee/analytics-warehouse/entities/ingestion-error.entity.ts
