+++ libs/ee/analytics-warehouse/entities/suggestion-mv.entity.ts
