+++ libs/ee/analytics-warehouse/modules/analytics-warehouse.module.ts
