+++ package.json
@@ -92,14 +92,14 @@
         "migration:revert:internal": "yarn typeorm migration:revert -d libs/core/infrastructure/database/typeorm/ormconfig.ts",
         "migration:run:prod": "node -r ./tsconfig-paths-bootstrap.js ./node_modules/typeorm/cli.js migration:run -d dist/libs/core/infrastructure/database/typeorm/ormconfig.js",
         "migration:revert:prod": "node -r ./tsconfig-paths-bootstrap.js ./node_modules/typeorm/cli.js migration:revert -d dist/libs/core/infrastructure/database/typeorm/ormconfig.js",
-        "analytics:migration:generate": "f() { API_PG_DB_HOST=localhost yarn typeorm migration:generate -p -d libs/analytics-warehouse/infrastructure/ormconfig.ts \"libs/analytics-warehouse/migrations/$@\"; }; f",
-        "analytics:migration:run": "API_PG_DB_HOST=localhost yarn typeorm migration:run -d libs/analytics-warehouse/infrastructure/ormconfig.ts",
-        "analytics:migration:revert": "API_PG_DB_HOST=localhost yarn typeorm migration:revert -d libs/analytics-warehouse/infrastructure/ormconfig.ts",
+        "analytics:migration:generate": "f() { API_PG_DB_HOST=localhost yarn typeorm migration:generate -p -d libs/ee/analytics-warehouse/infrastructure/ormconfig.ts \"libs/ee/analytics-warehouse/migrations/$@\"; }; f",
+        "analytics:migration:run": "API_PG_DB_HOST=localhost yarn typeorm migration:run -d libs/ee/analytics-warehouse/infrastructure/ormconfig.ts",
+        "analytics:migration:revert": "API_PG_DB_HOST=localhost yarn typeorm migration:revert -d libs/ee/analytics-warehouse/infrastructure/ormconfig.ts",
         "analytics:ensure-schema": "TS_NODE_PROJECT=tsconfig.json ts-node -r tsconfig-paths/register scripts/analytics/ensure-schema.cli.ts",
         "analytics:ensure-schema:prod": "node -r ./tsconfig-paths-bootstrap.js dist/scripts/analytics/ensure-schema.cli.js",
-        "analytics:migration:run:internal": "yarn typeorm migration:run -d libs/analytics-warehouse/infrastructure/ormconfig.ts",
-        "analytics:migration:revert:internal": "yarn typeorm migration:revert -d libs/analytics-warehouse/infrastructure/ormconfig.ts",
-        "analytics:migration:run:prod": "node -r ./tsconfig-paths-bootstrap.js ./node_modules/typeorm/cli.js migration:run -d dist/libs/analytics-warehouse/infrastructure/ormconfig.js",
+        "analytics:migration:run:internal": "yarn typeorm migration:run -d libs/ee/analytics-warehouse/infrastructure/ormconfig.ts",
+        "analytics:migration:revert:internal": "yarn typeorm migration:revert -d libs/ee/analytics-warehouse/infrastructure/ormconfig.ts",
+        "analytics:migration:run:prod": "node -r ./tsconfig-paths-bootstrap.js ./node_modules/typeorm/cli.js migration:run -d dist/libs/ee/analytics-warehouse/infrastructure/ormconfig.js",
         "analytics:backfill": "TS_NODE_PROJECT=tsconfig.json ts-node -r tsconfig-paths/register scripts/analytics/backfill.cli.ts",
         "analytics:backfill:prod": "node dist/apps/analytics-cli/main.js",
         "analytics:parity-check": "TS_NODE_PROJECT=tsconfig.json ts-node -r tsconfig-paths/register scripts/analytics/parity-check.ts",
